requirements.txt
numpy
pandas
yfinance
scikit-learn
ta
matplotlib
seaborn
streamlit
pyyaml
joblib
nltk
vaderSentiment
tensorflow
transformers
torch        
